requirements.txt
Flask==2.3.2
werkzeug>=2.1.0
transformers==4.28.1
torch==2.0.0
tensorflow==2.12.0
keras==2.12.0
emoji==1.7.0
python-dotenv==0.21.0
google-api-python-client==2.70.0
requests==2.28.1
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
matplotlib==3.7.1
jupyter==1.0.0
langdetect==1.0.9
googletrans==4.0.0-rc1
spacy==3.5.0
nltk==3.8.1
opendatasets==0.1.22
contractions==0.1.73
datasets==2.14.4
flask_cors==5.0.1